+++ src/Fields/Email.php
@@ -1,12 +1,12 @@
 <?php
 
-/*
- * This file is part of WordPlate.
+/**
+ * Copyright (c) Vincent Klaiber.
  *
- * (c) Vincent Klaiber <hello@doubledip.se>
- *
- * For the full copyright and license information; use please view the LICENSE
+ * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
+ *
+ * @see https://github.com/wordplate/extended-acf
  */
 
 declare(strict_types=1);