+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added `readonly` and `disabled` to date fields
 - Rename `Radio` class to `RadioButton`
 - Rename `Wysiwyg` class to `WysiwygEditor`
-- Updated field label to name conversion from kebab-case to snake_case
+- Updated field label to name conversion from `kebab-case` to `snake_case`
 
 ## 10.0.0